backend/requirements.txt
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# AI SDKs
anthropic==0.43.0
openai==1.59.9
google-genai==1.1.0
mistralai==1.3.1
cohere==5.13.3

# HTTP client (pour Ollama)
httpx==0.28.1

# GCP
google-cloud-firestore==2.19.0

# Utilitaires
python-dotenv==1.0.1

# Tests
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
